Prevent FOUC during fetch of next feed
This is done by turning the subscriptions#next action into a GET
request. The downside is that images that have not been entirely loaded
will continue to be fetched after the page change, even if they are not
to be showed anymore. I'll need to find some way of halting image load
in these situations.

References for FOUC:
http://foundation.zurb.com/sites/docs/dropdown-menu.html#preventing-fouc
https://en.wikipedia.org/wiki/Flash_of_unstyled_content